% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmlt_gof.R
\name{clean_capitalize}
\alias{clean_capitalize}
\title{Clean and Capitalize Variable Names}
\usage{
clean_capitalize(input_string)
}
\arguments{
\item{input_string}{Character. String to clean and capitalize.}
}
\value{
Character. Formatted string.
}
\description{
Internal helper function to format variable names for display by replacing
underscores and periods with spaces and capitalizing each word.
}
\keyword{internal}
